refactor: 병합후 파라미터값 수정
