+++ src/main/java/nbc/devmountain/domain/chat/chatroom/repository/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package nbc.devmountain.domain.chat.model.chatroom.repository;
+package nbc.devmountain.domain.chat.chatroom.repository;
 
 import java.util.List;
 
@@ -7,10 +7,11 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import nbc.devmountain.domain.chat.model.ChatRoom;
-import nbc.devmountain.domain.user.model.User;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom,Long> {
 
-	@Query("SELECT c FROM ChatRoom c WHERE c.user = :user AND c.deletedAt IS NULL")
-	List<ChatRoom> findAllByUser(@Param("user") User user);
+	@Query("SELECT c FROM ChatRoom c WHERE c.user.userId = :userId AND c.deletedAt IS NULL")
+	List<ChatRoom> findAllByUserId(@Param("userId") Long userId);
+
+
 }
